BREAKING_CHANGE Added support for Helm 3 environments
A Helm3Connector has been added to support the new version of Helm.
Pay attention that this is currently a breaking change, since now
the 'helm' connector type defaults to the 'helm3' connector.
In order to use the legacy 'helm2' connector, it is necessary to
explicitly specify 'helm2' type in the StreamFlow file