Added support for in-place dependency reading
When a dependency is already present in the worker node and the
workflow tasks are not going to modify the date, it is useful to
stage (copy) the entire file or directory into the staging dir.
Moreover, for huge datasets, this can be unfeasible. A simple
symlink function has been implemented to solve this issue