Added Kuberntes inCluster execution for Helm tempates
When StreamFlow runs inside a Kubernetes cluster, it can be useful
to rely on ServiceAccount objects to execute tasks directly on the
parent cluster, avoiding to explicitly include a Kubernetes config
file into the container. This commit solves such problem by adding
a dedicated inCluster option to Helm models configuration.