Created custom workflow management system
Instead of totally relying on cwltool for the execution of CWL workflow,
a custom StreamFlow workflow executor has been created. cwltool is still
used to parse the *.cwl files and to interpret the inline JavaScript code
according to the standard. The new structure makes it easy to extend the
StreamFlow features and to integrate with different coordination languages.

Some actions are still required to provide full CWL Standard adherence
(e.g. DockerRequirements, secondary files management, scatterMethods etc.).
These will be soon addressed with future updates to the software.