Added STMC support
Single Task Multiple Container (STMC) has been added to the StreamFlow
features by explicitly addressing launcher-based SPMD libraries such as
MPI. The `resources` directive allows to specify the number of resources
to allocate for a task. The related command will be executed only on the
first resource, with a STREAMFLOW_HOSTS environment variable populated
with the hostnames or IP addresses of all the reserved resources.

Regarding the data model, StreamFlow will copy the input data into all the
reserved resources (if necessary), but will look for the output files only
on the resource which actually execcuted the command. This is compatible
with both master-worker executions, where the master will aggregate the
data on its own node, or with shared data spaces scenarios, where all
the involved nodes share a portion of the file-system in which output
data will be generated.