Added SSH and Slurm connectors
In order to allow for remote execution on HPC infrastructures, two
new connectors have been added to the StreamFlow codebase:
  - An SSH connector, which can be used to connect with any generic
    SSH-enabled machine
  - A Slurm connector, which inherits from the SSH connector and
    extends it with `sbatch` submissions
The `from` configuration has been added to the `model` object to allow
for inheritance in environment configurations. Moreover, support for
Jinja2 templates has been added to allow the creation of customisable
scripts to launch command in the two new connectors.