Improved termination of workflow
Handling exceptions is really complicated in asyncio environments and
not feasible in distributed scenarios. Therefore, termination is now
handled directly by propagaitng termination tokens in the workflow
DAG, forcing termination of dependent jobs. This eliminates the need
for both a centralised termination handler in the executor and a
loop-level exception manager.