Improved remote execution performances
The need for an helper file for executing remote commands has been
removed, reducing to one third the amount of remote connections
needed to perform a command.

Moreover, the Kubernetes client now forces the `keep-alive` value
iuto the Connection header to prevent timeouts during the execution
of long remote tasks.