Added caching for available resources
The get-available_resources method is on the StreamFlow's critical
path, because it is called inside a synchronous code portion guarded
by a lock. Therefore, it must be as efficient as possible to avoid
performance bottlenecks.

The `cachetools` library has been introduced to cahce the results
for a configurable TTL before calling the remote resource retrieval
method again. This modification strongly enhances performances for
workflow with a large number of trivial tasks.