Control number of sessions per SSH connection
OpenSSH allows SSH servers to limit the number of concurrent SSH
sessions that can be opened from a single SSH connection. If more
sessions are opened, they are automatically discarded. A parameter
has been added to the SSH Connector to control this behaviour. The
default value is 10, according to the OpenSSH default.