Improved data tranfer logic
As an optimisation of the DataManager logic, now each input dependency
is transferred only once to each remote resource.

This is managed using the path mapper as a source of truth, and creating
symlinks to a primary location of an input dependency, even if it has
been transferred by a different, independent job.