Added secondary files propagation
Secondary files must be transferred alongside their primary entity
when moving computation on a remote model. Nevertheless, if secondary
files are specified when the transfer has already been performed, they
will not be present on the remote file-system and the command will fail.

To prevent this, secondary files are back-propagated toward the entire
workflow graph at buidl time, in order to ensure a correct behaviour.
This strategy introduces a (probably excessive) overhead, because
secondary files are processed even when it would not be necessary.
Smarter strategies can be found in the future to mitigate such overhead.