Added support for CWL conditional steps
The CWL 1.2+ standard supports conditional step execution through the
`when` directive. Since this directive can be applied to either a single
step or an entire subworkflow, a new NondeterminateMergeOutputCombinator
object has been introduced to handle multiple, alternative links in an
output port.

This enables the creation of skip links that bypass an entire subworkflow,
propagating a null token directly to the output port of the next step.