Get rid ov pipenv toolchain
The pipenv tool was initially adopted to guarantee repeatable builds
of the StreamFlow software. Nevertheless, it requires to manually
synchronize dependencies in both the setup.py and the Pipfile, which is
an useless complication.

From now on, StreamFlow will rely on the much more standard setuptools
build toolchain for both Docker and bare-metal installations.