Optimised dependencies in nested workflows
The standard way to manage step dependencies in CWL is to wait for
alll inputs to be ready before propagating them. When a step is a
nested workflow, this can introduce unnecessary barriers, e.g. forcing
the WMS to wait for a dependency that is not needed by the first step of
the workflow.

StreamFlow improved this mechanism with two additional features:
1) A js interpreter that can inspect step expressions and determine the
actual set of inputs for each step of the nested workflow;
2) A new translation mechanism that better regroups step inputs to
remove unnecessary constraints.