Added step-to-step bindings to the StreamFlow file
DAGs cannot easily deal with steps co-allocations, as this implies
a circular dependency between multiple steps. StreamFlow can overcome
this limitation with step-to-step bindings: two steps belonging to the
same group are always co-scheduled by the system, so that they can
communicate with each other if needed. This step is necessary to support
the Multiple-Tasks Multiple-Containers approach.