Added CWL conformance badges
This commit adds CWL Conformance badges to the README.md file and
substitutes references to Travis CI with the newly adopted GitHub
Actions toolchain.

This commit also introduces a resource-based scheduling for SSH nodes,
adding the logic for cores and memory inspection in the node. Disk
inspection is still inaccurate, as it checks only the root disk. A
correct inspection would require to modify the API.