Fixed docker GitHub Actions
An incorrect interpretation of the image existence environment variable
prevented the GitHub Actions workflow from creating the Docker images
for the new StreamFlow releases.

In particular, the returncode 0 was interpreted as a false value,
reversing the correct behaviour of the workflow.

In addition, the Helm version build arg was missing, preventing the
container build step from completing successfully.