Substituted `model` term with `deployment`
The term `model` is too generic and abused in workflow literature.
Therefore, it has been substituted with the term `deployment` both in
the StreamFlow config schema and in the source code.

The old term `model` has been deprecated and warning messages have been
added in the logs to notify users. It will be removed in StreamFlow
0.3.0.