Fixed Helm Connector, apsw and Windows execution
This commit fixes two errors in the Helm connector:
 - The function checking if helm CLI is installed on the local machine
was failing due to an error in the conditional clause. This error is
completely fixed;
 - The function checking the Helm version fails because of an error in
the uvloop implementation (issue 457). This error is still unsolved, but
the uvloop has been substituted with the standard asyncio
implementation. If the error will be solved, uvloop will be used again
due to its superior performances.

The official releases of the apsw libraries are not released on PyPI.
This makes it harder to install the requirement. For this reason,
StreamFlow now relies on the standard sqlite3 Python client.

Compatibility with Windows was broken because the local connector was
using the `sh` shell to run commands. A proper check that relies on
`cmd` on Windows has been added.