Improved supoport for Windows platform
Windows is not supported anymore by cwltool, as it only supports WSL2.
This commit addresses some of the StreamFlow issues with Windows, such as
substituting PosixPath with PurePosixPath, removing calls to unsupported os
functions, and retrieving the HOME path in a platform-agnostic way.

The road to Windows compatibility is still difficult, mainly due to the
incompatibility of cwltool JavaScript libraries that are called everywhere
in the code. Waiting for a potential workaround, StreamFlow drops the support
for the Windows platform for the moment.