Moved most log messages to DEBUG level
Complex workflows with hundreds or thousands of micro-steps generated a
huge amount of logs, which made it diffucult for users to focus on
effective workflow outputs. Most of these log messages have been moved
from the INFO log level to the DEBUG log level. Plus, the profiling
utilities have been removed to clean the log output at the DEBUG level.