Created the aiotarstream library for async tar
In order to improve data transfer performances, a new interface for
asynchronous tar streams has been created. This interface wraps
asynchronous streams behind a common interface (the StreamWrapper class)
and lets the tar library handle the creation and extraction of archives
in a fully asynchronous way, without storing the intermediate data on
the disk or fully in memory.