Added services for SSH and QueueManager connectors
Before this commit, it was possible to specify only a single `file`
template for each SSH and QueueManager deployment. However, it is better
to distinguish between functional specification (contained in the `file`
template) and the system-level specification (contained in other
properties, e.g., the connection string or the maxConcurrentJobs).

This commit allows to configure multiple functional interfaces for the
same deployment, unifying the behaviour of low-level SSH and
QueueManager connectors to the more cloud-oriented ones.