Improving performance by removing remote commands
The `exists` and `follow_symlink` commands can be fused in a single
command that returns `None` when the remote file does not exist. The
same is valid also for the `checksum` command.

The `follow_symlink` remote command can be avoided when at least one
PRIMARY location is already registered on the data manager, as it
contains the real path referenced by the current (potential) symlink.

The `checksum` command is redundant when the data transfer results in a
symbolic link, as the dst file cannot but be equal to the original.

The presence of explicit `port` targets makes it redundant to check for
the presence of a file/directory in the remote location.