Update CWL toolchain compatibility
Recent commits in the CWL toolchain moved some functionalities from the
cwltool runner to the more general cwl-utils codebase. This commit
updates StreamFlow to be compatible with the new APIs.

Plus, this commit freezes the versions of the included CWL components,
in order to avoid unwanted incompatibilities with future updates.