Added multi-target bindings
This commit introduces multi-target bindings for steps. Each step of a
workflow can target a list of deployments and services, and it will be
scheduled on one of them depending on the runtime availability of
resources.

In addition, a list of `filters` can be configured to manipulate the
list of available targets, e.g. to change the order of availability
evaluation or to exclude some of them. By default, targets are evaluated
in the order they are declared in the `binding` configuration.