Check returncode when submit job to queue manager
This commit fixes a wrong behaviour of the QueueManager connector, which
failed when the `_get_returncode()` function did not receive an `int`
value from the queue call. This happens when the job submission itself
fails. This fix explicitly captures the return code of the queue
manager invocation and throws an exception when receiving an error code.

Fixes #18.