Added support for nested connectors
This commit adds support for nested Connector classes through the
`ConnectorWrapper` interface. The `QueueManagerConnector` is now
implemented as a connector wrapper, automatically creating an inner SSH
connector to maintain compatibility with the (deprecated) previous
version, which inherited from the `SSHConnector` class.

A new `wraps` property in the `deployment` configuration allows to
define connector wrappers for other connectors. A single connector can
be wrapped from multiple connector wrappers, but each wrapper can wrap
a single inner connector.

The `deploy` and `undeploy` actions are not automatically propagated to
the inner connector to avoid spurious creations and deletions of
connectors when multiple wrappers wrap the same object. Instead, the
correct order of deployment and undeployment actions is in charge of the
`DeploymentManager` class.