Avoid sqlite race conditions during tests
Parallel CWL conformance tests often fail due to a `Database is locked`
exception thrown by Sqlite. Indeed, concurrent access to a single Sqlite
file is not well supported by the library.

This commit uses a dedicated in-memory Sqlite instance for each test, in
order to speed up the overall execution and to avoid race conditions.