Fixed MacOS X compatibility (#77)
This commit fixes support for MacOS X operating system. In particular it
fixes the following bugs:
 - On MacOS X, temporary directory can be a symbolic link. This commit
explicitly derives the real path when creating a temporary folder;
 - On MacOS X, the default version of sh is quite old and doesn't support
all command features. This commit explicitly invokes bash shell on Darwin
systems;
 - The CWL translator was mounting the host StreamFlow workdir with a bind
mount to the same path in a Docker container. Now the folder in the container
is always configured as `/tmp/streamflow`.
In addition, this commit sets up a Mac OS X runner to perform automatic CI
unit tests and CWL conformance tests when submitting PRs.