Updated and fixed StreamFlow Dockerfile (#97)
This commit fixes an error on the Dockerfile specification, which
prevented the StreamFlow dependencies from being correctly installed.
Plus, it moves streamflow from the `/root/.local/` folder to the
`/opt/streamflow` system-wide location, allowing non-root executions of
the container. Finally, it updates the root container from Python3.8 to
Python 3.11. Fixes #94.