Fix Docker and Helm features
Now that StreamFlow defines multiple subcommands other than `run`, the
`ENTRYPOINT` has been removed from the Dockerfile, allowing users to run
any command they like inside the container.

A test has been added to the CI pipeline to check that the Docker image
is working properly.

The Helm Chart has also been modified accordingly. The Job command was not
working correctly, because it contained a spurious `selector` entry. Now
it has been fixed, and it defaults to a simple job that returns the
StreamFlow version.