Fix type checking in conftest and flux Dockerfile (#195)
Type comparison should always be done using either `isinstance` or
`is` and `is not` operators. This commit fixes a wrong behaviour in
the `conftest.py` file.

This commit also explicitly installs the `attrs` Python package in the
Flux Docker image, to prevent a frequent `module not found` error in the
CI pipeline.