Fix Flux and Mac OS CI steps (#209)
The Flux CI step often fails because the `attrs` module cannot be found
on the system. This commit fixes this behaviour by explicitly installing
the module through `pip` in the CI workflow.

As per issue lima-vm/lima#1742, the QEMU 8.1.0 Homebrew package is
broken on Intel architectures. This commit upgrades the
`douglascamata/setup-docker-macos-action` to version `v1-alpha.9` to
overcome this issue.