Fix scheduler behaviour when retry_delay is set (#224)
This commit addresses two problems with the `retry_delay` option of the
`DefaultScheduler` class. First, it ensures that there is no retry when
the `retry_delay` option is set to 0 (as reported in the docs). Second,
it captures and handles the `TimeoutError` raised when the `retry_delay`
timeout is reached.