Fix potential race condition on sqlite `row_factory`
Internally, the `SqliteDatabase` used a dirty trick to switch between
standard `row_factory` (which returns a tuple) and the `aiosqlite.Row`
object (which returns a dict). This commit forces all methods to return
a dict object, removing a potential race condition that sometimes caused
the methods of `SqliteDatabase` to return the wrong data type.