Avoid opening too many SSH sessions
Sometimes, a race condition caused the `SSHConnector` to open more than
the maximum allowed sessions (specified through the
`maxConcurrentSessions` parameter). This led to broken connections and
errors in workflow executions.

This commit prvents this error by refactoring the `SSHConnector`
internals. It also adds a no-regression test that transfers 20 files
simultaneously to and from a remote SSH location.