Removed deprecated `pkg_resources` API
The `pkg_resources` API has been deprecated in favor of the new
`importlib_resources` API. This commit replaces all the usages of
`pkg_resources` with the new API.

To implement this new path, the `get_schema` method of each StreamFlow
extension must now return the content of the schema instead of the path
to the JSON Schema file. Documentation has been updated accordingly.