Add `optional` param to composite processors
Composite `TokenProcessor` objects failed to validate `Token` value
when receiving `None`, even if `optional` was set to true. Indeed, the
`optional` parameter was only set to the inner `processor` objects,
which were called after a first `Token` validation.

This commit explicitly sets the `optional` parameter to the outer
`TokenProcessor` object, solving the issue.