Fixed Config persistence (#289)
This commit fixes the `ScheduleStep` persistence, particularly the `BindingFilter` object.
The `filters` attribute in `BindingFilter` was not being saved correctly in the database.
A new class extends the `Config`, called `FilterConfig`, and a new table was added to the database.

The extention of `PersistableEntity` was moved from `Config` to its superclasses only where needed, i.e. `DeploymentConfig` and `FilterConfig`.