Adjusted local-to-remote copy with folders
The local-to-remote copy was failing when trying to copy a folder with
multiple files, as only the first file was actually tansferred. This
commit solves this issue by modifying the `aiotarstream` module to deal
with `tell` and `seek` operations, allowing to position the stream
pointer in the correct position.

In detail, two new classes have been implemented:

- `TellableStreamWrapper`, which keeps track of the actual stream
   pointer position;
- `SeekableStreamReaderWrapper`, which allows forward seek operations
  in a stream reader.

Note that these classes are internal to the `aiotarstram` module, so
nothing changes in the publicly exposed API.