Fixed remote-to-local copy of nested directories (#309)
This commit fixes the remote-to-local copy of nested directories. When transferring
nested directory to a remote location, an empty directory was created with the same
name of her parent. This depended on a wrong handling on paths in the
`base Connector`. To avoid regressions, a new unit test was added.

Furthermore, this commit fixes the `seek` coroutine in the `aiotarstream` module,
which was never awaited, and correctly handles the case when a `seek` is asked
to move the stream pointer to its current position, which resulted in a wrongly-raised
exception.