Fix `valueFrom` type checking behaviour
This commit fixes the type checking behaviour with CWL `valueFrom`
transformers. Before, the incorrect type was passed to the
`TokenProcessor` class: the type resulting from the `valueFrom`
expression evaluation. Instead, the correct type should be the type of
the `source` value, i.e., before the expression evaluation. Since this
type is difficult to infer, as the source could be still unknown at the
evaluation time, now StreamFlow always passes `Any`.