Fix job name (#325)
This commit fixes the uniqueness of job names. Before this commit,
the job name was composed of the step name and the last tag level
of its inputs. However, only the last tag level is not enough because
some jobs could have the same name. In particular, nested scatter
ports can generate tags such as 0.0.0 and 0.1.0, thus there
will be two jobs with the same name step_name/0. Now, the
full tag is used to identfy jobs.