Fixed symbolic link remote input data error (#328)
This commit fixes the case when input remote data involve a symbolic link
path. The path was saved in the StreamFlow filesystem as a regular file,
without any check. Now, the right file type is saved (`PRIMARY` for
regular files, `SYMBOLIC_LINK` for symbolic links). 
Plus, for symbolic links, also the primary path is saved.

Futhermore, this commit fixes a race condition in remote data transfers.
If a remote data object is required by two concurrent steps on the same
location and the two steps are executed in parallel, the first one starts
copying the object, while the second creates a symbolic link from the data
copied by the first one. However, the second must wait for the first one
to end the copy operation before creating the symbolic link.

Finally, this commit adds new in the logger to notify the end of copy
operations.