Added check after the creation data (#336)
This commit adds a new check when the data are created.
This check helps avoiding wrong creations.

In some cases it is possible that the data are not found, e.g.,
when a location fails and the data have been lost, or the user
sets a wrong path in the configuration file.

Before this commit, the data was created empty in the new
location. Now, an exception is thrown.