Reduce size of database metadata (#338)
This commit reduces the footprint of database metadata
by removing the `output` column of the `command` table,
which stored the `stdout` of a `Job execution`. Indeed, the
output of a command can be costly in terms of data size,
and up to now was not used by StreamFlow.

Futhermore, the `command` table has been renamed to
avoid confusion with the StreamFlow `Command` class.
Indeed, the `Command` class is not a `PersistableEntity`,
but is saved inside the `step` database table.