Add `__slots__` attribute to `Token` objects (#385)
This commit optimizes the StreamFlow memory footprint by adding the
`__slots__` attribute to all `Token` objects. Indeed, if the `__slots__`
attribute is not added to a subclass (even when empty), that class is
built with the standard `__dict__` field.

Plus, this commit adds the `__slots__` field to the `PersistableEntity`
class, because otherwise all classes inheriting from it will have the
`__dict__` field, too. Since Python does not support multiple
inheritance when two or more classes define the `__slots__` attribute,
this commit also refactors the `DeploymentConfig` and `FilterConfig`
classes to inherit only from `PersistableEntity`.