Fixed `asyncio.Event` synchronizations when an exception occurs (#395)
This commit introduces a mechanism to awake pending executions
on the `asyncio.Event`. When there are concurrent tasks that want
to deploy a location, only one task makes the deployment, while the
others wait for the first to terminate. However, if the first gets an
exception, all the other tasks must be awakened and raise a 
`WorkflowExecutionException`.