Fix `ListMergeCombinator` port names (#407)
This commit fixes #400. In general, the local port names in StreamFlow have a
POSIX-like name, but they must not be absolute paths. Instead, the step
names are absolute paths. In the case of `ListMergeCombinator`, the local port
names were absolute paths, causing the issues reported in #400.

In addition, the collector step name was changed with the absolute path following
the style defined before. Furthermore a typo of a variable name in the
`_get_deploy_step` method was fixed.