Fixed glob on symbolic link path (#417)
This commit fixes an issue when checking glob on a step output that involves
a symbolic link path. Before this commit, StreamFlow was just checking if the
`dirname` of the output real path was the same of some `Job` directories.
However, if `Job` directories were containing a symbolic link path the check
failed.

This commit ensures that all the directories stored in a `Job` object are
real paths, without symbolic links. In this way, the glob checking never
involves symbolic links, fixing the issue.