Fixed workdir of wrapped deployments (#461)
This commit fixes the inheritance mechanism for the `workdir` property of wrapped deployments. In the case where the wrapped deployment has not defined a `workdir`, but its wrapper deployment does, the first inherits from the second.

Moreover, the `_inner_deploy` method of the `DefaultDeploymentManager` class now injects the `workdir` into the `DeploymentConfig` of the inner connector. The previous behaviour did not lead to errors, but it was corrected to avoid future ones.