Check job status when it terminates (#452)
This commit introduces a check of the job status inside the `run`
method of the `ExecuteStep` class. It is helpful to terminate other
running jobs of the same step, because if a job exits with an error
status, the entire workflow run must fail. Such a failure does not
depend on the manager, as the failure is handled inside the `_run_job`
method itself.

Before this commit, the step was checking the job status only when all
the jobs terminated their executions (either successfully or not). However,
such a behavior results in a waste of time and computing resources.