Fix remote write (#471)
This commit fixes the creation of a file in a remote location.
Before this commit, some special characters (e.g., the dollar
sign `$`) were not escaped correctly. To avoid this problem, the
file content is now encoded locally in base64 and decoded in the
remote location, redirecting the `stdout` to the target file.