Fixed output step when it is an empty list (#472)
This commit returns an empty list when no data are generated
from a CWL step that expects an `array` output type. In addition,
this commit forces all output processors to be `optional` when an
`array` output type is specified, as the CWL standard allows falling
back to an empty list even for non-optional output parameters.