Implementing stacked locations scheduling
When a `ConnectorWrapper` wraps an inner `Connector` object, the
locations returned by the `get_available_locations` method could either
be allocated directly on the lower hardware stack (e.g., Docker
containers) or be offloaded to remote machines (e.g., the jobs of a
Slurm allocation).

This commit introduces the `stacked` flag in the `AvailableLocation`
object to specify if a location represents the former (`true`) or the
latter (`false`, the default value) case.

In addition, this commit instructs the `Scheduler` to take into account
the whole hardware stack when validating resource requirements, slightly
modifying how the `LocationAllocation` objects are populated.