This commit reduces the memory space required by a workflow instance, particularly in the database and also in RAM when the workflow is loaded from the database.
Before this commit, when the workflow created, each `CWLTokenProcessor` instance shared the same `Graph` instance. However, each `CWLTokenProcessor` instance saved in the database their `Graph` instance independently. It duplicated the same data, and it was highly costly. Moreover, when a new `Workflow` was loaded from the database, each `CWLTokenProcessor` instance created their own `Graph` instance because there was no logic that made them share the same instance.

Now, the `CWLWorkflow` has an attribute to store the `Graph` instance. And each `CWLTokenProcessor` refers to it when needed.