Fixed `CONTENT_LIMIT` checks (#491)
This commit fixes the `CONTENT_LIMIT` check when `loadContent`
feature is enabled. Before this commit, an exception was always
raised when the `CONTENT_LIMIT` was not being respected in the
output file. Instead, in the CWL v1.0 and v1.1, no errors should be raised,
but only the first `CONTENT_LIMIT` bytes should be taken.

Moreover, this commit fixes an erroneous propagation of the `load_content`
parameter when the output ports have already been processed by
the `outputEval` clause

Finally, this commit fixes the `CWLTransferStep` persistence, correctly
loading the `writable` attribute. Related unit tests have been improved
to check this case.