Added explicit `return` in `_copy_remote_to_remote` (#515)
This commit adds explicit `return` statements in all the branches
of the `if-else` construct in the `_copy_remote_to_remote` method
of the `ContainerConnector` class. Before this commit, only one branch
of the `if-else` construct had the `return` statement. This fix is necessary
to be complaint to the `py/mixed-returns` rule of the `CodeQL` tool.