Moved `Graph` reference from `CWLTokenProcessor` to `CWLWorkflow` (#516)
This commit reduces the memory space required by a workflow
instance, particularly in the database and also in RAM when the
workflow is loaded from the database.

Before this commit, when a `Workflow` was instantiated, each
`CWLTokenProcessor` instance shared the same `Graph` instance.
However, each `CWLTokenProcessor` instance saved in the database
their own `Graph` replica independently. This behaviour was highly
costly and generated duplicated data. Moreover, when a new `Workflow`
was loaded from the database, each `CWLTokenProcessor` instance
created its own `Graph` instance because there was no logic that made
them share the same instance.

Now, the `CWLWorkflow` has an attribute to store the `Graph` instance,
and each `CWLTokenProcessor` refers to it when needed.