Changed propagation of failure termination (#373)
This commit changes how StreamFlow propagates the failure
termination into the workflow steps. Before this commit, the steps
were propagating a `SKIPPED` status because the ports were empty.
This commit uses the 'TerminationToken' value to propagate the
`Step` exit status. This way, the following steps know if they can continue
the execution or terminate immediately.