Improved description about location storage in the `Hardware` class (#423)
This commit enriches the location storage information in the `Hardware`
class. Before this commit, the input, output and temp directories of a `Job`
were managed as independent entities; instead, they should be correlated
when they are located in the same persistent volume. For example, if a step
requires 16 GB in the output directory and 8 GB in the temp directory and
the directories are in the same volume, the volume needs to have at least
24 GB of available storage.

Now, the attribute `storage` in the `Hardware` class maps each persistent
volume with a new object called `Storage`. The `Storage` class describes
the volume with a `mount_point` attribute, i.e., the path where the volume
is mounted in the related location. A `size` attribute contains the available
space in the volume, while a `paths` dictionary contains the set of visited
paths belonging to the target volume. 

---------

Co-authored-by: GlassOfWhiskey <iacopo.c92@gmail.com>