Fixed job input for the CWL `InitialWorkDirRequirement`
This commit fixes the job inputs after the evaluation of the changes for
the CWL `InitialWorkDirRequirement` . Before this commit, the data was
copied in the job working directory, but the paths inside the job `inputs`
attribute were not update. This can bring to some issues when in the
CWL inputs are used the full input path.