Fix `follow_symlink` function return value
This commit fixes the `follow_symlink` function. Before this commit,
when the path did not exist, the local case returned `None`, while
the remote case returned an empty string.

Moreover, this commit fixes the `get_mount_point` function, since it
internally calls the `follow_symlink`. After this commit, the ad-hoc fix
for the `follow_symlink` is not needed anymore.