Handle lists in CWL `glob` field
According to the CWL standard specification, the `glob` field of an
`outputBinding` section can be an array of strings. In that case, a
compliant WMS should find files or directories that match any pattern in
the array. This case was not handled by StreamFlow, and this commit
implements support for it.