Fixed optional CWL parameter (#466)
This commit fixes the optional parameter feature. Before this commit,
empty `Token` objects were injected into ports without input steps.
However, an execution deadlock occurred when `scatter` or `loop`
features of CWL were included in a workflow.

After this commit, optional parameters are handled as null `default`
values. Specifically, for each input parameter that is not explicitly linked
to the output of a previous step, a `DefaultTransformer` with the `None`
default value is injected into the workflow representation. Whether a
parameter is used or not can be detected while translating a workflow
from CWL to StreamFlow dataflow semantics. Indeed, if a the parameter
is present in an inner CWL object (`CommandLineTool`, `ExpressionTool`,
or nested `Workflow`) but not in the wrapping `WorkflowStep` object, and
the parameter type is optional, the fix can be applied.