Improved `SSHConnector` error message (#574)
This commit supplies a more informative message error. The first
command executed on the `SSHConnector` retrievals the hardware
information. However, whenever the remote location prints an error
message to `stderr`, an error is raised without a comprehensive
explanation. Now, an exception with an exhaustive error message is raised. 

Moreover, the logging of commands executed remotely though
`SSHConnector` were not printed because there was a wrong association
between an inline if-else and two f-strings.