Fixed `InputInjectorStep` in remote location data case (#566)
This commit fixes an issue introduced in #423 . This fix allows to
have initial workflow data located in a remote location. In the
StreamFlow file, it is possible to bind the ports with a location and
define the `workdir` of the ports. The workflow config file (e.g. the
CWL config file) defines the relative path based on the `workdir`
defined in the StreamFlow file.

Added a non-regression test on the correct injection of the data
with the remote path.