Fix concurrency in `DefaultScheduler`
Now that hardware resources are checked for thw whole stack of
`Connector` objects, just locking on a single `Condition` related to the
outermost `Connector` is not enough to prevent race conditions. This
commit ensures that a `Lock` object for each `Connector` in the stack is
acquired prior to reserve or release resources during scheduling. Locks
are always acquired and released in order, preventing deadlocks.

In addition, this commit makes the `ConnectorWrapper` compatible with
the `FutureConnector` feature by extending the `FutureAware` class.