Fixed `workdir` inheritance in wrapped deployments (#588)
This commit fixes the retrieval of `workdir` when a `wraps` option
has a `deployment` and `service` defined in the StreamFlow file. Before
this commit, an exception was raised when the `BindingConfig` object
was created because StreamFlow was assuming that `wraps` was a string. 

Moreover, this commit fixes the inheritance scheme of the `workdir`
option. Before this commit, `workdir` was retrieved only from the first
wrapped  deployment. Now, multiple levels of inheritance are correctly
handled. The topmost defined `workdir` in the `Connector` hierarchy is
assigned to the `BindingConfig` instance. Some unit tests have been added
to check the correct inheritance of the  `workdir` attribute when `wraps`
is defined.